refactor: Move experiment/scan to server
This better oragnizes the handling of the api representation of an
application to an in cluster representation of an application.
